feat: cli tool done
need to be tested on project
